Modularize branch selector surfaces
