+++ src/app/components/workspace/composer/composer-diff-baseline-selector.tsx
@@ -1,34 +1,23 @@
 import { useQuery } from '@tanstack/react-query'
-import { Check, GitCompareArrows, GitPullRequestDraft, Search } from 'lucide-react'
+import { GitCompareArrows } from 'lucide-react'
 import { type RefObject, useEffect, useId, useMemo, useRef, useState } from 'react'
-import { createPortal } from 'react-dom'
 import type {
   ProjectCommitEntry,
   ProjectDiffBaseline,
   ProjectDiffStatsResult,
   ProjectGitState,
 } from '../../../desktop/types'
-import {
-  getFeatureStatusBadgeClass,
-  getFeatureStatusDataAttributes,
-} from '../../../features/feature-status'
 import { useDismissibleLayer } from '../../../hooks/useDismissibleLayer'
 import {
   desktopQueryKeys,
   getProjectDiffStatsQuery,
   listProjectCommitsQuery,
 } from '../../../query/desktop-query'
-import { popoverPanelClass, settingsInputClass } from '../../../ui/classes'
 import { cn } from '../../../utils/cn'
-import { SurfacePanel } from '../../common/surface-panel'
-import {
-  BaselineOption,
-  baselineOptions,
-  CommitOption,
-  getBaselineCounts,
-  matchesCommitSearch,
-} from './composer-diff-baseline-options'
+import { getBaselineCounts, matchesCommitSearch } from './composer-diff-baseline-options'
 import { getDiffBaselineLabel, getDiffBaselinePrefix } from './diff-baseline'
+import { BaselineSelectorPortal } from './diff-baseline-selector/baseline-selector-popover'
+import { ComposerBranchSelectorPopover } from './diff-baseline-selector/branch-selector-popover'
 import { formatGitCount } from './git-ops'
 import {
   type BaselineAnchorKind,
@@ -277,26 +266,6 @@ function BaselineBranchButton({
   open: boolean
   panelId: string
 }) {
-  const openBranchSwitch = () => {
-    if (!onSwitchBranch) {
-      onOpen()
-      return
-    }
-
-    onOpen()
-  }
-
-  const submitBranchSwitch = () => {
-    const nextBranch = branchSwitchInput.trim()
-    if (!(nextBranch && onSwitchBranch)) return
-    onSwitchBranch(nextBranch)
-    onSetBranchSwitchOpen(false)
-  }
-  const filteredBranches = branches.filter((branch) =>
-    branch.toLowerCase().includes(branchSwitchInput.trim().toLowerCase()),
-  )
-  const visibleBranches = filteredBranches.slice(0, 5)
-
   return (
     <span className="relative inline-flex">
       <button
@@ -313,211 +282,29 @@ function BaselineBranchButton({
         onPointerDownCapture={(event) => {
           event.preventDefault()
           event.stopPropagation()
-          openBranchSwitch()
+          onOpen()
         }}
-        onClick={openBranchSwitch}
+        onClick={onOpen}
         onMouseEnter={onPreview}
       >
         <span className="truncate">{branchLabel}</span>
       </button>
-      {branchSwitchOpen && typeof document !== 'undefined'
-        ? createPortal(
-            <SurfacePanel
-              id={panelId}
-              ref={branchSwitchPanelRef}
-              data-open={positionReady ? 'true' : 'false'}
-              aria-label="Branch selector"
-              className={cn(
-                popoverPanelClass,
-                'motion-popover fixed z-[160] grid max-h-[calc(100vh-1rem)] grid-rows-[auto_auto_minmax(0,1fr)_auto] gap-2 rounded-2xl p-2 transition-[opacity,transform] duration-150 ease-out',
-                positionReady
-                  ? 'translate-y-0 opacity-100'
-                  : 'pointer-events-none translate-y-1 opacity-0',
-              )}
-              style={{
-                bottom: `${panelPosition.bottom}px`,
-                left: `${panelPosition.left}px`,
-                maxHeight: `${panelPosition.maxHeight}px`,
-                width: `${panelPosition.width}px`,
-              }}
-            >
-              <div className="px-1 text-[11px] uppercase tracking-[0.08em] text-[color:var(--muted)]">
-                Switch branch
-              </div>
-              <div className="grid min-h-0 gap-0.5 overflow-y-auto pb-0.5">
-                {visibleBranches.length > 0 ? (
-                  visibleBranches.map((branch) => (
-                    <button
-                      key={branch}
-                      type="button"
-                      className={cn(
-                        'grid min-h-9 w-full grid-cols-[16px_minmax(0,1fr)] items-center gap-2 rounded-xl px-2.5 py-2 text-left text-[12.5px] text-[color:var(--muted)] transition-colors hover:bg-[rgba(255,255,255,0.04)] hover:text-[color:var(--text)]',
-                        branch === branchLabel &&
-                          'bg-[rgba(255,255,255,0.06)] text-[color:var(--text)]',
-                      )}
-                      onClick={() => {
-                        if (!onSwitchBranch) return
-                        onSwitchBranch(branch)
-                        onSetBranchSwitchOpen(false)
-                      }}
-                    >
-                      <span className="inline-flex items-center justify-center text-[color:var(--accent)]">
-                        {branch === branchLabel ? <Check size={13} /> : null}
-                      </span>
-                      <span className="truncate">{branch}</span>
-                    </button>
-                  ))
-                ) : (
-                  <div className="px-2 py-1.5 text-[12px] text-[color:var(--muted)]">
-                    Press Enter to check out “{branchSwitchInput.trim()}”
-                  </div>
-                )}
-              </div>
-              <input
-                ref={branchSwitchInputRef}
-                value={branchSwitchInput}
-                onChange={(event) => onSetBranchSwitchInput(event.target.value)}
-                onKeyDown={(event) => {
-                  if (event.key === 'Enter') submitBranchSwitch()
-                  if (event.key === 'Escape') onSetBranchSwitchOpen(false)
-                }}
-                className={settingsInputClass}
-                placeholder="Search branches"
-              />
-              <div
-                className="grid min-h-[92px] gap-2 rounded-xl border border-dashed border-[color:var(--border)] bg-[rgba(255,255,255,0.025)] p-3"
-                {...getFeatureStatusDataAttributes('feature:composer.worktrees')}
-              >
-                <div className="flex items-center justify-between gap-3">
-                  <div className="flex min-w-0 items-center gap-2">
-                    <span className="inline-flex h-7 w-7 shrink-0 items-center justify-center rounded-lg border border-[color:var(--border)] bg-[color:var(--panel-2)] text-[color:var(--muted)]">
-                      <GitPullRequestDraft size={14} />
-                    </span>
-                    <div className="min-w-0">
-                      <div className="truncate text-[12.5px] font-medium text-[color:var(--text)]">
-                        Worktrees
-                      </div>
-                      <div className="truncate text-[11px] text-[color:var(--muted)]">
-                        Reserved for linked branch workspaces.
-                      </div>
-                    </div>
-                  </div>
-                  <span className={getFeatureStatusBadgeClass('feature:composer.worktrees')}>
-                    Mock
-                  </span>
-                </div>
-                <div className="grid grid-cols-2 gap-2 text-[11px] text-[color:var(--muted)]">
-                  <div className="rounded-lg border border-[color:var(--border)] bg-[rgba(255,255,255,0.02)] px-2 py-1.5">
-                    Create worktree
-                  </div>
-                  <div className="rounded-lg border border-[color:var(--border)] bg-[rgba(255,255,255,0.02)] px-2 py-1.5">
-                    Open existing
-                  </div>
-                </div>
-              </div>
-            </SurfacePanel>,
-            document.body,
-          )
-        : null}
+      {branchSwitchOpen && onSwitchBranch ? (
+        <ComposerBranchSelectorPopover
+          branchLabel={branchLabel}
+          branches={branches}
+          branchSwitchInput={branchSwitchInput}
+          inputRef={branchSwitchInputRef}
+          panelId={panelId}
+          panelPosition={panelPosition}
+          panelRef={branchSwitchPanelRef}
+          positionReady={positionReady}
+          onSetBranchSwitchInput={onSetBranchSwitchInput}
+          onSetBranchSwitchOpen={onSetBranchSwitchOpen}
+          onSwitchBranch={onSwitchBranch}
+        />
+      ) : null}
     </span>
-  )
-}
-
-function BaselineSelectorPortal({
-  commitsQuery,
-  panelId,
-  panelPosition,
-  panelRef,
-  positionReady,
-  searchQuery,
-  selectedBaseline,
-  selectedCommitSha,
-  setOpen,
-  setSearchQuery,
-  visibleCommits,
-  onSelectBaseline,
-}: {
-  commitsQuery: ReturnType<typeof useQuery<ProjectCommitEntry[]>>
-  onSelectBaseline: (baseline: ProjectDiffBaseline) => void
-  panelId: string
-  panelPosition: { left: number; bottom: number; width: number; maxHeight: number }
-  panelRef: RefObject<HTMLDivElement | null>
-  positionReady: boolean
-  searchQuery: string
-  selectedBaseline: ProjectDiffBaseline
-  selectedCommitSha: string | null
-  setOpen: (open: boolean) => void
-  setSearchQuery: (query: string) => void
-  visibleCommits: ProjectCommitEntry[]
-}) {
-  if (typeof document === 'undefined') return null
-  const panelLeft = `${panelPosition.left}px`
-  const panelWidth = `${panelPosition.width}px`
-  return createPortal(
-    <SurfacePanel
-      id={panelId}
-      ref={panelRef}
-      data-open={positionReady ? 'true' : 'false'}
-      aria-label="Diff baseline selector"
-      className={cn(
-        popoverPanelClass,
-        'motion-popover fixed z-[120] grid max-h-[calc(100vh-1rem)] grid-rows-[auto_minmax(0,1fr)_auto_auto] gap-2 rounded-2xl p-2 transition-[opacity,transform] duration-150 ease-out',
-        positionReady ? 'opacity-100 translate-y-0' : 'pointer-events-none opacity-0 translate-y-1',
-      )}
-      style={{
-        bottom: `${panelPosition.bottom}px`,
-        left: panelLeft,
-        maxHeight: `${panelPosition.maxHeight}px`,
-        width: panelWidth,
-      }}
-    >
-      <div className="px-2 pt-1 text-[11px] uppercase tracking-[0.08em] text-[color:var(--muted)]">
-        Changes since
-      </div>
-      <div className="grid min-h-0 gap-0.5 overflow-y-auto pb-0.5">
-        {visibleCommits.length > 0 ? (
-          visibleCommits.map((commit) => (
-            <CommitOption
-              key={commit.sha}
-              commit={commit}
-              selected={selectedCommitSha === commit.sha}
-              onSelect={() => {
-                onSelectBaseline({ kind: 'commit', sha: commit.sha })
-                setOpen(false)
-              }}
-            />
-          ))
-        ) : (
-          <div className="px-2.5 py-3 text-[12px] text-[color:var(--muted)]">
-            {commitsQuery.isLoading ? 'Loading commits…' : 'No commits found.'}
-          </div>
-        )}
-      </div>
-      <label className="relative block">
-        <Search
-          size={14}
-          className="pointer-events-none absolute top-1/2 left-3 -translate-y-1/2 text-[color:var(--muted)]"
-        />
-        <input
-          value={searchQuery}
-          onChange={(event) => setSearchQuery(event.target.value)}
-          placeholder="Search commits"
-          className={cn(settingsInputClass, 'w-full pl-9')}
-        />
-      </label>
-      {baselineOptions.map((option) => (
-        <BaselineOption
-          key={option.key}
-          label={option.label}
-          selected={selectedBaseline.kind === option.key}
-          onSelect={() => {
-            onSelectBaseline(option.baseline)
-            setOpen(false)
-          }}
-        />
-      ))}
-    </SurfacePanel>,
-    document.body,
   )
 }
